+++ NEWS.md
@@ -1,5 +1,7 @@
+
 # dv.tables 0.1.0
 
+* Add Tplyr_table module 
 * Add jumping functionality
 * Implement label truncation
 * Implement functionality to download to Word (.rtf) format